Created the DrawPong class and refactored RunPongGame to simplify the implementation
